LSP-Liskov Substitution Principle
The Liskov Substitution Principle (LSP) is one of the five SOLID principles of object-oriented programming. It states that objects of a superclass should be replaceable with objects of a subclass without affecting the correctness of the program. In other words, a subclass should be able to be used interchangeably with its parent class without causing unexpected behavior.